tesselle: Easily Install and Load 'tesselle' Packages

Easy install and load key packages from the 'tesselle' suite in a single step. The 'tesselle' suite is a collection of packages for research and teaching in archaeology. These packages focus on quantitative analysis methods developed for archaeology. The 'tesselle' packages are designed to work seamlessly together and to complement general-purpose and other specialized statistical packages. These packages can be used to explore and analyze common data types in archaeology: count data, compositional data and chronological data. Learn more about 'tesselle' at https://www.tesselle.org.

Details

Package:	tesselle

Type:	Package

Version:	1.5.0

License:	GPL-3

See also

Author

Full list of authors and contributors (alphabetic order):

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

Archéosciences Bordeaux (UMR 6034)
Maison de l'Archéologie
Université Bordeaux Montaigne
33607 Pessac cedex
France